PiecewiseLinearSpectrum & use acs
