+++ .gitignore
@@ -12,3 +12,7 @@ Cargo.lock
 
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
+
+# Appearance color space tables
+*.acsc
+*.acss
