+++ Cargo.toml
@@ -24,6 +24,7 @@ appearance-wgpu = { version = "0.1.0", path = "crates/appearance-wgpu" }
 appearance-world = { version = "0.1.0", path = "crates/appearance-world" }
 
 anyhow = { version = "1.0.94", default-features = false, features = ["std"] }
+appearance-color-spaces = { git = "https://github.com/TemporalInteractive/appearance-color-spaces.git", rev = "7a31bc0", default-features = false }
 bytemuck = { version = "1.13", default-features = false, features = ["derive", "extern_crate_alloc", "min_const_generics"] }
 clap = { version = "4.5.27", default-features = true, features = ["derive"] }
 # egui = { version = "0.30.0", default-features = true, features = ["bytemuck"] }
